Util: Update minidns to 1.1.1
from https://github.com/MiniDNS/minidns tag 1.1.1 2024-10-24

Possible fixes (unverified):
- MiniDNS per-RR parsers throw NegativeArraySizeException on malformed header
- DnsName.parse recurses per label without depth bound StackOverflow DoS
- DnsName parser throws AIOOBE (not IOException) on truncated pointer
- ipv4From / ipv6From resolve via DNS instead of parsing literals
- Question constructor uses assert for null checks
- DnsMessage pre-allocates ArrayLists from 16-bit header counts
- OPT.parse uses dis.read() and disabled assert for bounds

reported by: bottomlineit.co.za

+++ core/java/src/org/minidns/record/NSEC.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2015-2020 the original author or authors
+ * Copyright 2015-2024 the original author or authors
  *
  * This software is licensed under the Apache License, Version 2.0,
  * the GNU Lesser General Public License version 2 or later ("LGPL")
@@ -110,7 +110,7 @@ public class NSEC extends Data {
                 }
                 int a = (type >> 3) % 32;
                 int b = type % 8;
-                bitmap[a] |= 128 >> b;
+                bitmap[a] |= (byte) (128 >> b);
             }
             if (windowBlock != -1) writeOutBlock(bitmap, dos);
         } catch (IOException e) {
